move char codec to additional file
